Modify BERT dependency with add_pooler control
